requirements.txt
APScheduler>=3.2.0
click>=6.6
Flask>=0.11.1
Jinja2>=2.8
pynamodb>=1.5.3
validator.py>=1.2.5